Set up zsh with p10k
Sets recommended font to xterm and makes it use 256 colors, necessary
for the rainbow prompt to show on "p10k configure".

Sets "dotfiles" alias for the git bare repository of my dotfiles.

.zprofile file for setting base environment variables, and defining the
location of the zsh config files. Still needs to execute "startx" on
"/dev/tty1".

The setting of the PATH variable is being done in .zshrc, that at the
moment only adds the "$HOME/.local/bin" directory.