Set monitor specific settings and changing to feh
All the previous commits where on a Virtual Machine, now that I'm in a
real machine I can be more specific with the monitor settings.

Since the laptop's montitor is broken, I disable it. And use instead an
external monitor for it.

Also this commit changes restoring pywal's colorscheme, and instead uses
feh to only set the wallpaper. This is for having a consistent
colorcheme.