Update sxhkd config
Separates the base bindings from the custom ones, more to test if I
can save myself some complete keyboard freezes, that I need to restart
the machine to fix, that happens when I make a mistake on its syntax.
Dunno if this is a fix to this atm tho.

The terminal binding uses the environment variable TERMINAL defined in
the ~/.zprofile file.
Removes all style arguments from the dmenu_run binding, since it gets
it from Xresources anyway.