Change theme to MaterialOcean
Polybar configuration has also been changed to use only the colors on
the theme, that are less than the ones that were previously.

Bspwm uses some of the colors present on the ~/.Xresources file for the
windows, with the help of a Python script. I tried to use the "xrdb -q"
command and pipe it, but it didn't actually change the colors at
startup.

Oh, and dmenu gets to use colors of this theme too.