Change polybar bars
Got in the mood to change it. Now I have more space than with the old
one, allowing me to add more info to the network module and still have
room for something else.

Saw around a lot of people placing the workspaces in the center of the
bar, so I want to try it out.

Also, using backgrounds for the modules instead of underlines. For now
doesn't seem bad to me. ~~Let's see how much this lasts.~~