Add birthday script and polybar module
The module calls as bash script to not have the crazy line for opening
the list in vim with st. This way at least is easier to change the
location from the script and csv file.

Also, dunno where to place it. Wanted to do it inside ~/.config/polybar
but it just didn't felt right to me, at least right now. So it is where
it is at the moment.