Cleaning home a little
Huh, so ~/.zshenv is called every time it seems. ~~Better learning that
later than never.~~

Haven't seen a way to make ~/.fehbg be created on another place, so I
stick with the wallpaper restore system that I have with the addition
of an environment variable to control where the wall file is stored.

Wanted to move ~/.powerlevel10k to $XDG_DATA_DIR/zsh but got a not
readable error, so it is where it ended up at the moment.

Most probably I forgot something.