Move start menu startup to `zprofile`
I have the necessity of populating `PATH` before the start menu itself
runs, some things go weirdly otherwise iirc

Recently got a solution, just move the PATH inserting bit to another
file and source it where needed while keeping a single point of error. I
feel stupid not thinking of this then.

The early exit on the start-menu script is a little of an unnecessary
fail-safe, there is no need for it.